+++ src/bot.js
@@ -10,14 +10,6 @@ const { MessageEmbed } = require("discord.js");
 
 const client = new WelcomeBot();
 
-process.env.userAgent = "Discord Welcome-Bot " + client.botVersion;
-process.on("unhandledRejection", (error) => {
-    console.error("Unhandled promise rejection:", error);
-});
-process.on("exit", (code) => {
-    client.destroy();
-});
-
 const presence = require("./functions/presence");
 const greetUser = require("./functions/greetUser");
 const sayGoodBye = require("./functions/sayGoodBye");
@@ -29,6 +21,21 @@ const addGuild = require("./db/functions/guild/addGuild");
 const removeGuild = require("./db/functions/guild/removeGuild");
 const getGuild = require("./db/functions/guild/getGuild");
 const dbAuditor = require("./db/functions/dbAuditor");
+
+process.env.userAgent = "Discord Welcome-Bot " + client.botVersion;
+process.on("unhandledRejection", (error) => {
+    console.error("Unhandled promise rejection:", error);
+    if (
+        error.message.indexOf("No guild with guild ID") !== -1 &&
+        client &&
+        dbAuditor
+    ) {
+        dbAuditor(client);
+    }
+});
+process.on("exit", (code) => {
+    client.destroy();
+});
 
 client.on("ready", () => {
     // We logged in
@@ -67,13 +74,19 @@ client.on("guildMemberRemove", (member) => {
 client.on("guildCreate", (guild) => {
     //Bot has been invited to a new guild
     addGuild(guild.id);
-    guild.channels.cache
-        .find((ch) => ch.id === guild.systemChannelID)
-        .send("Thank you for choosing this bot! To get started, type `w/help`");
+    if (guild.systemChannelID) {
+        guild.channels.cache
+            .get(guild.systemChannelID)
+            .send(
+                "Thank you for choosing this bot! To get started, type `w/help`"
+            );
+    }
     let embed = new MessageEmbed()
         .setTitle(`Added to "${guild.name}"`)
         .setDescription(`${guild.id}`);
-    client.channels.cache.get(client.loggingChannelId).send(embed);
+    client.channels.cache
+        .get(client.loggingChannelId)
+        .send({ embeds: [embed] });
 });
 
 client.on("guildDelete", (guild) => {
@@ -82,7 +95,9 @@ client.on("guildDelete", (guild) => {
     let embed = new MessageEmbed()
         .setTitle(`Added to "${guild.name}"`)
         .setDescription(`${guild.id}`);
-    client.channels.cache.get(client.loggingChannelId).send(embed);
+    client.channels.cache
+        .get(client.loggingChannelId)
+        .send({ embeds: [embed] });
 });
 
 client.on("message", async function (message) {