+++ src/commands/info/user.js
@@ -9,7 +9,6 @@ module.exports = {
     description:
         "Get information about a user. It will show your info if no user was mentioned",
     args: false,
-    guildOnly: true,
     usage: "(@mention / user_id)",
     category: "Information",
     async execute(message, args) {
@@ -36,9 +35,12 @@ module.exports = {
         if (!user) {
             return false;
         }
-        let member = message.guild.members.cache.find((m) => m.id === user.id);
-        if (!member) {
-            return message.reply("That user was not found in this server");
+        let member;
+        if (message.guild) {
+            member = message.guild.members.cache.find((m) => m.id === user.id);
+            if (!member) {
+                return message.reply("That user was not found in this server");
+            }
         }
 
         /*let badges = [];
@@ -75,10 +77,14 @@ module.exports = {
         }*/
         embed.addField(
             "Joined:",
-            `Joined discord at *${user.createdAt}*\n\nJoined **${message.guild.name}** server at *${message.member.joinedAt}*`
+            `Joined discord at *${user.createdAt}*` +
+                (message.guild
+                    ? `\n\nJoined **${message.guild.name}** server at *${message.member.joinedAt}*`
+                    : "")
         );
         embed.addField("Locale:", `${user.locale}`);
-        if (member.nickname) embed.addField("Nickname:", `${member.Nickname}`);
+        if (member && member.nickname)
+            embed.addField("Nickname:", `${member.Nickname}`);
         //https://discord.js.org/#/docs/main/stable/class/User?scrollTo=presence
         embed.addField("Presence:", `${user.presence.status}`);
         embed.setFooter(